:recycle: refactor: ajustar gráfico de barras de evento por quantidade
